update
[title="First Real Update"]
I got a good idea of what what going on with the structure of a system
like this based around babylon.js so now it is time to start deploying a
real working model.  The old prototype is in an archive folder and
should still work, I have not tested it yet though.  Updates and release
should be coming soon.